tests/submission/Ms2ger/obsolete-features/requirements-for-implementations/other-elements-attributes-and-apis/document.all-01.html

<!DOCTYPE html>
<title>document.all: willfull violations</title>
<link rel="author" title="Ms2ger" href="mailto:ms2ger@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#dom-document-all">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_equals(Boolean(document.all), false);
  assert_equals(document.all == undefined, true,
    "(document.all == undefined) isn't true");
  assert_equals(document.all != undefined, false,
    "(document.all != undefined) isn't false");
  assert_equals(document.all === undefined, false,
    "(document.all === undefined) isn't false");
  assert_equals(document.all !== undefined, true,
    "(document.all !== undefined) isn't true");
  assert_equals(typeof document.all, "undefined");
  assert_equals("all" in document, true, "\"all\" in document returns false");
});
</script>
